requirements.txt
agency-swarm>=1.0.1
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0